distbuild: Set log file size limit
The 'log-keep' setting defaults to 10, so systems will contain a max 1GB
of log data. Previously they would have eventually used up all of the
available space in /var with log data.

We may want to set 'log-level=info' as a default for future deployments,
but for now it's useful being able to do detailed analysis of errors.