Add write extension for making a package that allows deferred instansiation
This is a write extension for making a package that can be used to
install the produced system. Ideally we'd instead have Baserock
everywhere to do the deployment, but we need to support this workflow
until that is possible.

It gets configured with a script templates that get filled in with the
paths to the rootfs tarball as @@ROOTFS_TAR_PATH@@, script directory as
@@SCRIPT_DIR@@ and copied boot-files as @@IMAGE_DIR@@.

Scripts to include are specified as a relative path to your morphologies
repository, separated by : characters in the INCLUDE_SCRIPTS variable.
If another separator is convenient, it can be specified in the
INCLUDE_SCRIPTS_SEPARATOR variable, which has the same semantics as the
shell IFS variable.

Other files needed to create the disk image, such as bootloader image
files can be specified in BOOTLOADER_BLOBS, relative to the root of the
rootfs, separated with similar rules to INCLUDE_SCRIPTS.

The name of the rootfs tarball, the scripts subdirectory and the
bootloader blobs subdirectory can be changed with the ROOTFS_TAR,
SCRIPT_SUBDIR and IMAGE_SUBDIR variables, but the default should be
sufficient.

The compression of the rootfs tarball, and the package tarball can be
specified by specifying a shell command that reads the uncompressed
tarball from stdin and writes to stdout in the ROOTFS_COMPRESS and
OUTPUT_COMPRESS variables.