Update all references between morphologies to be filenames
When the chunk morphologies are moved into definitions, morphologies
will be referred to by their filename rather than by the morphology
name. This allows us to simplify morph edit, as well as be specific when
referring to chunks from strata, meaning that the directory structure
does not affect the functionality of morph.