Improve cloud-init.configure extension to detect cloud-init
Some users think that adding the cloud-init configuration
extension to a system is enough to install and enable cloud-init.

This is not true, to add and enable cloud-init in a system the
system has to have cloud-init installed (cloud-init chunk) and
cloud-init services have to be enabled (with cloud-init
configuration extension)

This patch is to alert the user that cloud-init can't be enabled
in a system without cloud-init installed, and making the
deployment fail.